modified:   .metadata/.lock_info 	modified:   .metadata/.log 	modified:   .metadata/.mylyn/.tasks.xml.zip 	modified:   .metadata/.mylyn/tasks.xml.zip 	modified:   .metadata/.plugins/org.eclipse.core.resources/.safetable/org.eclipse.core.resources 	modified:   .metadata/.plugins/org.eclipse.e4.workbench/workbench.xmi 	modified:   .metadata/.plugins/org.eclipse.php.ui/pdt-images/0.png 	modified:   .metadata/.plugins/org.eclipse.ui.workbench.texteditor/dialog_settings.xml 	modified:   .metadata/.plugins/org.eclipse.wst.validation/dep.index 	modified:   .metadata/version.ini 	modified:   biblioteca/mvc/controllers/EjemplarController.php 	modified:   biblioteca/mvc/controllers/LibroController.php 	modified:   biblioteca/mvc/controllers/PanelController.php 	modified:   biblioteca/mvc/controllers/PrestamoController.php 	modified:   biblioteca/mvc/controllers/SocioController.php 	modified:   biblioteca/mvc/controllers/TemaController.php 	modified:   biblioteca/mvc/controllers/UserController.php 	modified:   biblioteca/mvc/views/libro/list.php 	modified:   biblioteca/mvc/views/libro/show.php 	modified:   biblioteca/mvc/views/panel/admin.php 	modified:   biblioteca/mvc/views/panel/panel.php 	modified:   biblioteca/mvc/views/socio/create.php 	modified:   biblioteca/mvc/views/socio/delete.php 	modified:   biblioteca/mvc/views/socio/edit.php 	modified:   biblioteca/mvc/views/socio/show.php 	modified:   biblioteca/mvc/views/tema/create.php 	modified:   biblioteca/mvc/views/tema/delete.php 	modified:   biblioteca/mvc/views/tema/edit.php 	modified:   biblioteca/mvc/views/tema/list.php 	modified:   biblioteca/mvc/views/tema/show.php 	new file:   biblioteca/public/images/users/user_67cf73cd425919.64123818.png

+++ biblioteca/mvc/controllers/EjemplarController.php
@@ -1,5 +1,5 @@
 <?php
-use http\Message;
+
 
 /**
  * EjemplarController
@@ -19,7 +19,10 @@ class EjemplarController extends Controller{
 	 * @return ViewResponse
 	 */
 	public function index(){
-		return $this->list();
+		if( Login::role('ROLE_LIBRARIAN' ))// autorización(solo bibliotecarios)
+				return $this->list();
+		//Si no es bibliotecario, redirige a la home
+		return redirect('/');
 	}
 	
 	/**
@@ -30,11 +33,14 @@ class EjemplarController extends Controller{
 	 */
 	public function list(){
 		
-		
-		$ejemplares= Ejemplar::all(); // recupera los ejemplares del libro
-		
-		//	carga la vista que los muestra
-		return view('ejemplar/list',['ejemplares'=>$ejemplares]);
+		if( Login::role('ROLE_LIBRARIAN' )) {// autorización(solo bibliotecarios)
+			
+			$ejemplares= Ejemplar::all(); // recupera los ejemplares del libro
+			//	carga la vista que los muestra
+			return view('ejemplar/list',['ejemplares'=>$ejemplares]);
+		}
+		//Si no es bibliotecario, redirige a la home
+		return redirect('/');
 	}
 	
 	/**
@@ -44,15 +50,18 @@ class EjemplarController extends Controller{
 	 */
 	public function show(int $id=0) {
 		
-		
-		// Recupera el libro
-		$libro = Ejemplar::findOrFail($id, 'No se encontró el ejemplar indicado'); //tb comprueba si no le ha llegado el ID
-		
-		//recupera los ejemplares del libro
-		$ejemplares= $libro->hasMany('Ejemplar');
-		
-		// carga la vista y le pasa el libro recuperado
-		return view ('ejemplar/show',['libro'=>$libro,'ejemplares'=>$ejemplares]);
+	  	if( Login::role('ROLE_LIBRARIAN' )) {// autorización(solo bibliotecarios)
+			// Recupera el libro
+			$libro = Ejemplar::findOrFail($id, 'No se encontró el ejemplar indicado'); //tb comprueba si no le ha llegado el ID
+			
+			//recupera los ejemplares del libro
+			$ejemplares= $libro->hasMany('Ejemplar');
+			
+			// carga la vista y le pasa el libro recuperado
+			return view ('ejemplar/show',['libro'=>$libro,'ejemplares'=>$ejemplares]);
+	  	}
+	  	//Si no es bibliotecario, redirige a la home
+	  	return redirect('/');
 		
 	}
 	
@@ -62,10 +71,16 @@ class EjemplarController extends Controller{
 	 */
 	public function create(int $idlibro=-1){
 		
-		$libro = Libro::findOrFail($idlibro,'No se encontró el libro.');
+		if( Login::role('ROLE_LIBRARIAN' )) {// autorización(solo bibliotecarios)
+			
+			$libro = Libro::findOrFail($idlibro,'No se encontró el libro.');
 		
-		//retorna una ViewResponse con la vista con el formulario de creacion
-		return view('Ejemplar/create',['libro'=>$libro]);
+			//retorna una ViewResponse con la vista con el formulario de creacion
+			return view('Ejemplar/create',['libro'=>$libro]);
+		}
+		//Si no es bibliotecario, redirige a la home
+		return redirect('/');
+
 	}
 	
 	/**
@@ -74,15 +89,18 @@ class EjemplarController extends Controller{
 	 * @ redirect Viewresponse
 	 */
 	public function store(){
-		//Comprueba que la petición venga del formulario
-		if(!request()->has('guardar'))
-			throw new FormException('No se recibió el formulario');
-		$ejemplar=new Ejemplar(); //crea el nuevo ejemplar
-		
-	//OPCION AUTOMATICA
+		if( Login::role('ROLE_LIBRARIAN' )) {// autorización(solo bibliotecarios)
+			
+			//Comprueba que la petición venga del formulario
+			if(!request()->has('guardar'))
+				throw new FormException('No se recibió el formulario');
+			$ejemplar=new Ejemplar(); //crea el nuevo ejemplar
+			
+		//OPCION AUTOMATICA
 			try{
 				
 				//guarda el libro en la base de datos a partir de los datosPOST
+				$ejemplar->saneate(); //sanea las entradas.
 				$ejemplar = Ejemplar::create(request()->posts()); 
 				
 				
@@ -109,6 +127,10 @@ class EjemplarController extends Controller{
 				//regresa al formulario de creación de libro
 				return redirect("/Ejemplar/create/$libro->id");
 			}
+		}
+		//Si no es bibliotecario, redirige a la home
+		return redirect('/');
+		
 
 	}
 	
@@ -120,44 +142,51 @@ class EjemplarController extends Controller{
 	 * @return ViewResponse
 	 */
 	public function delete(int $id=0){
-		
-		$libro = Libro::findOrFail($id, "No existe el libro.");
-		
-		return view('libro/delete',['libro'=> $libro]);
+		if( Login::role('ROLE_LIBRARIAN' )) {// autorización(solo bibliotecarios)
+			$libro = Libro::findOrFail($id, "No existe el libro.");
+			
+			return view('libro/delete',['libro'=> $libro]);
+		}
+		//Si no es bibliotecario, redirige a la home
+		return redirect('/');
+	
 	}
 	
 	/** Elimina el ejemplar de la base de datos
 	 * @return RedirectResponse
 	 */
 	public function destroy(int $id=-1){
-
-		//Recupera el elemplar de la BDD			
-		$ejemplar	=Ejemplar::findOrFail($id, "No se encontró el ejemplar.");
-		$libro=Libro::findOrFail($ejemplar->idlibro,"No se ha encontrado el libro");
-		//Si hay prestamos, no permitimos el borrado
-		try{
-			if($ejemplar->hasAny('Prestamo','idejemplar')){
-				throw new Exception("No se puede borrar el ejemplar mientras está prestado.");
+		if( Login::role('ROLE_LIBRARIAN' )) {// autorización(solo bibliotecarios)
+			//Recupera el elemplar de la BDD			
+			$ejemplar	=Ejemplar::findOrFail($id, "No se encontró el ejemplar.");
+			$libro=Libro::findOrFail($ejemplar->idlibro,"No se ha encontrado el libro");
+			//Si hay prestamos, no permitimos el borrado
+			try{
+				if($ejemplar->hasAny('Prestamo','idejemplar')){
+					throw new Exception("No se puede borrar el ejemplar mientras está prestado.");
+					return redirect("/Libro/edit/$ejemplar->idlibro");
+				}
+			} catch (Exception $e){
+							Session::error($e->getMessage());
+				
+					return redirect("/Libro/edit/$ejemplar->idlibro");
+			}	
+				//intenta borrar el ejemplar
+			try{
+					$ejemplar->deleteObject();
+					Session::success("Se ha borrado el ejemplar $ejemplar->id de libro $libro->titulo.");
+					return redirect("/Libro/edit/$ejemplar->idlibro");
+					//si se produce un error en la operació con la bdd..
+			} catch (Exception $e){
+				
+				Session::error("No se pudo borrar el ejemplar $ejemplar->id de  $libro->titulo.");
+				
+				
 				return redirect("/Libro/edit/$ejemplar->idlibro");
 			}
-		} catch (Exception $e){
-						Session::error($e->getMessage());
-			
-				return redirect("/Libro/edit/$ejemplar->idlibro");
-		}	
-			//intenta borrar el ejemplar
-		try{
-				$ejemplar->deleteObject();
-				Session::success("Se ha borrado el ejemplar $ejemplar->id de libro $libro->titulo.");
-				return redirect("/Libro/edit/$ejemplar->idlibro");
-				//si se produce un error en la operació con la bdd..
-		} catch (Exception $e){
-			
-			Session::error("No se pudo borrar el ejemplar $ejemplar->id de  $libro->titulo.");
-			
-			
-			return redirect("/Libro/edit/$ejemplar->idlibro");
 		}
+		//Si no es bibliotecario, redirige a la home
+		return redirect('/');
 		
 	}
 	
@@ -170,69 +199,76 @@ class EjemplarController extends Controller{
 	 *
 	 */
 	public function edit(int $id=0){
-		
-		// busca el libro con ese ID
-		$ejemplar = Ejemplar::findOrFail($id,'No se encontró el ejemplar.');
-		
-		//recupera los ejemplares del libro
-		 $libro= $ejemplar->belongsTo('Libro');
-		
-		//retorna una ViewResponse con la vista con el formulario de edición
-		 return view('ejemplar/edit',['ejemplar'=>$ejemplar, 'libro'=>$libro]);
+		if( Login::role('ROLE_LIBRARIAN' )) {// autorización(solo bibliotecarios)
+			// busca el libro con ese ID
+			$ejemplar = Ejemplar::findOrFail($id,'No se encontró el ejemplar.');
+			
+			//recupera los ejemplares del libro
+			 $libro= $ejemplar->belongsTo('Libro');
+			
+			//retorna una ViewResponse con la vista con el formulario de edición
+			 return view('ejemplar/edit',['ejemplar'=>$ejemplar, 'libro'=>$libro]);
+	
+		}
+		//Si no es bibliotecario, redirige a la home
+		return redirect('/');
 	}
 	
 	/** Actualzia la bdd con los datos POST del formulario
 	 */
 	public function update(){
-		
-		if(!request()->has('actualizar')) //si no llega el formulario ...
-			throw new FormException ('No se recibieron datos');
-			
-			$id = intval(request()->post('id')); // recuperar el id via POST
-			
-			
-			//Con la actualización a 1.8.0 ya se puede recuperar el formulario y tratarlo directamente
-			/*
-			 $libro = Libro::findOrFail($id,"No se ha encontrado el libro.");
-			 
-			 //recuperar el resto de campos
-			 $libro->isbn	= request()->post('isbn');
-			 $libro->titulo	= request()->post('isbn');
-			 $libro->editorial	= request()->post('isbn');
-			 $libro->autor	= request()->post('isbn');
-			 $libro->idioma	= request()->post('isbn');
-			 $libro->edicion	= request()->post('isbn');
-			 $libro->anyo	= request()->post('isbn');
-			 $libro->edadrecomendada	= request()->post('isbn');
-			 $libro->paginas	= request()->post('isbn');
-			 $libro->caracteristicas = request()->post('isbn');
-			 $libro->sinopsis	= request()->post('isbn');
-			 */
-			$libro= request()->post('idlibro');
-			//intenta actualizar el libro
-			try{
-				//$libro->update(); No es necesario en la 1.8.0
-				// ya el metodo create ya actualiza si manda el 2ºparametro
-				$ejemplar= Ejemplar::create(request()->posts() ,$id);
+		if( Login::role('ROLE_LIBRARIAN' )) {// autorización(solo bibliotecarios)
+			if(!request()->has('actualizar')) //si no llega el formulario ...
+				throw new FormException ('No se recibieron datos');
 				
-				Session::success("Actualización del ejemplar $ejemplar->id del libro $ejemplar->idlibro correcta.");
-				return redirect("/Libro/edit/$ejemplar->idlibro");
+				$id = intval(request()->post('id')); // recuperar el id via POST
 				
-				// Si se produce un error al guardar el libro..
-			}catch (SQLException $e){
-				// prepara el mensaje de error
-				$mensaje = "No se pudo actualizar el ejemplar";
 				
-				if(str_contains($e->errorMessage(),'Duplicate entry'))
-					$mensaje.="<br>Ya existe un ejemplar con ese <b>ID</b>.";
-					Session::error($mensaje);
+				//Con la actualización a 1.8.0 ya se puede recuperar el formulario y tratarlo directamente
+				/*
+				 $libro = Libro::findOrFail($id,"No se ha encontrado el libro.");
+				 
+				 //recuperar el resto de campos
+				 $libro->isbn	= request()->post('isbn');
+				 $libro->titulo	= request()->post('isbn');
+				 $libro->editorial	= request()->post('isbn');
+				 $libro->autor	= request()->post('isbn');
+				 $libro->idioma	= request()->post('isbn');
+				 $libro->edicion	= request()->post('isbn');
+				 $libro->anyo	= request()->post('isbn');
+				 $libro->edadrecomendada	= request()->post('isbn');
+				 $libro->paginas	= request()->post('isbn');
+				 $libro->caracteristicas = request()->post('isbn');
+				 $libro->sinopsis	= request()->post('isbn');
+				 */
+				$libro= request()->post('idlibro');
+				//intenta actualizar el libro
+				try{
+					//$libro->update(); No es necesario en la 1.8.0
+					// ya el metodo create ya actualiza si manda el 2ºparametro
+					$ejemplar->saneate(); //sanea las entradas.
+					$ejemplar= Ejemplar::create(request()->posts() ,$id);
 					
-					if(DEBUG)
-						throw new SQLException($e->getMessage());
+					Session::success("Actualización del ejemplar $ejemplar->id del libro $ejemplar->idlibro correcta.");
+					return redirect("/Libro/edit/$ejemplar->idlibro");
+					
+					// Si se produce un error al guardar el libro..
+				}catch (SQLException $e){
+					// prepara el mensaje de error
+					$mensaje = "No se pudo actualizar el ejemplar";
+					
+					if(str_contains($e->errorMessage(),'Duplicate entry'))
+						$mensaje.="<br>Ya existe un ejemplar con ese <b>ID</b>.";
+						Session::error($mensaje);
 						
-						return redirect("/Libro/edit/$ejemplar->idlibro");
+						if(DEBUG)
+							throw new SQLException($e->getMessage());
+							
+							return redirect("/Libro/edit/$ejemplar->idlibro");
+				}
 			}
+		
+		//Si no es bibliotecario, redirige a la home
+		return redirect('/');
 	}
-	
-	
 }
